+++ src/API/arabigo.js
@@ -1,0 +1,3 @@
+const { parse } = require('./api');
+
+console.log(parse('XX'));
